+++ sktime/classification/kernel_based/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+__all__ = [
+    "ROCKETClassifier",
+    "Arsenal",
+]
+
+from sktime.classification.kernel_based._arsenal import Arsenal
+from sktime.classification.kernel_based._rocket_classifier import ROCKETClassifier
