+++ sktime/classification/kernel_based/_arsenal.py
