+++ sktime/classification/kernel_based/_rocket_classifier.py
