+++ sktime/classification/shapelet_based/__init__.py
@@ -2,11 +2,7 @@
 __all__ = [
     "ShapeletTransformClassifier",
     "MrSEQLClassifier",
-    "ROCKETClassifier",
-    "Arsenal",
 ]
 
-from sktime.classification.shapelet_based._arsenal import Arsenal
 from sktime.classification.shapelet_based._stc import ShapeletTransformClassifier
 from sktime.classification.shapelet_based.mrseql.mrseql import MrSEQLClassifier
-from sktime.classification.shapelet_based._rocket_classifier import ROCKETClassifier
